+++ src/components/pages/profile/AddProfileForm.js
@@ -3,11 +3,11 @@ import ProfileInputs from "./ProfileInputs";
 
 function AddProfileForm({ handleSubmit, addHandleChange }) {
   return (
-    <></>
-        // <ProfileInputs
-        //   handleSubmit={handleSubmit}
-        //   addHandleChange={addHandleChange}
-        // />
+  
+        <ProfileInputs
+          handleSubmit={handleSubmit}
+          addHandleChange={addHandleChange}
+        />
   );
 }
 export default AddProfileForm;
