+++ src/components/pages/profile/Profile.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState, useContext } from "react";
 import NavBar from "../../Navbar.js";
-//import ProfileInfo from "./ProfileInfo";
+import ProfileInfo from "./ProfileInfo";
 import AddProfile from "./AddProfile.js";
 import EditUser from "./EditUser";
 import Container from "@material-ui/core/Container";
@@ -11,7 +11,7 @@ import RecommendedComponent from "../../recomended/RecommendedComponent";
 import Favorites from "../../Favorites/Favorites.js";
 import { createUserContext } from "../../../functions";
 import UserContext from "../../../contexts/UserContext";
-import TestForm from "../../../components/pages/profile/TestForm"
+
 
 export default function Profile(props) {
   const [editing, setEditing] = useState(false);
@@ -38,8 +38,8 @@ export default function Profile(props) {
       <Container maxWidth="lg">
         <Grid container spacing={3}>
           <Grid item xs={4}>
-            {/*<ProfileInfo toggleEditing={toggleEditing} />*/}
-            <TestForm/>
+            <ProfileInfo toggleEditing={toggleEditing} />
+            
           </Grid>
           <AddProfile profileData={profileData} />
           {editing ? (
